#37 fix ffq
