+++ .github/workflows/test.yaml
@@ -29,10 +29,8 @@ jobs:
 
       - name: Run tests
         run: |
-          chmod 111 testdata/utils/read_index/ffq/index_invalid_permission
-          go test -v -coverprofile=coverage.out $(go list ./... | grep -v "examples")
+          go test -v -shuffle=on -race -timeout 20s -coverprofile=coverage.out $(go list ./... | grep -v "examples")
           go tool cover -func=coverage.out
-          chmod 644 testdata/utils/read_index/ffq/index_invalid_permission
 
       - name: Upload test results
         if: always()
