+++ options_test.go
@@ -1,45 +1,40 @@
 package ffq
 
 import (
-	"strings"
+	"errors"
 	"testing"
 )
 
 func TestWithFileDir(t *testing.T) {
-	input := "/tmp"
-	var expectRet error = nil
-	expectedVal := input
-
+	dirName := "/tmp"
 	var options options
-	f := WithFileDir(input)
-	actualRet := f(&options)
-	if expectRet != actualRet {
-		t.Fatalf("Failed test: expectRet: %v, actualRet: %v", expectRet, actualRet)
+	f := WithFileDir(dirName)
+	err := f(&options)
+
+	if err != nil {
+		t.Errorf("failed test: got is nil, %v", err)
 	}
-	actualVal := options.fileDir
-	if *actualVal != expectedVal {
-		t.Fatalf("Failed test: expectedVal: %v, actualVal: %v", expectedVal, actualVal)
+
+	if *options.fileDir != dirName {
+		t.Errorf("failed test: got is equal dirName, %s, %s", *options.fileDir, dirName)
 	}
 }
 
 func TestWithQueueSize(t *testing.T) {
 	tests := []struct {
-		name        string
-		input       int
-		expectRet   string
-		expectedVal int
+		name  string
+		input uint64
+		err   error
 	}{
 		{
-			name:        "queue size can set",
-			input:       16,
-			expectRet:   "",
-			expectedVal: 16,
+			name:  "queue size can set",
+			input: 16,
+			err:   nil,
 		},
 		{
-			name:        "queue size cannot set with less than 1",
-			input:       0,
-			expectRet:   "queueSize must be set to greater than 0",
-			expectedVal: 0, // nil
+			name:  "queue size cannot set with less than 1",
+			input: 0,
+			err:   ErrQueueOption,
 		},
 	}
 
@@ -47,25 +42,12 @@ func TestWithQueueSize(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			var options options
 			f := WithQueueSize(tt.input)
-			actualRet := f(&options)
-			if actualRet == nil {
-				if tt.expectRet != "" {
-					t.Fatalf("Failed test: %s, expectRet: %v, actualRet: %v", tt.name, tt.expectRet, actualRet)
-				}
-			} else {
-				if !strings.Contains(actualRet.Error(), tt.expectRet) {
-					t.Fatalf("Failed test: %s, expectRet: %v, actualRet: %v", tt.name, tt.expectRet, actualRet)
-				}
+			err := f(&options)
+			if !errors.Is(err, tt.err) {
+				t.Errorf("failed test: got is not equal err, %v, %v", err, tt.err)
 			}
-			actualVal := options.queueSize
-			if actualVal == nil {
-				if tt.expectedVal != 0 {
-					t.Fatalf("Failed test: %s, expectedVal: %v, actualVal: %v", tt.name, tt.expectedVal, actualVal)
-				}
-			} else {
-				if *actualVal != tt.expectedVal {
-					t.Fatalf("Failed test: %s, expectedVal: %v, actualVal: %v", tt.name, tt.expectedVal, actualVal)
-				}
+			if err == nil && *options.size != tt.input {
+				t.Errorf("failed test: got is nil, %v", err)
 			}
 		})
 	}
@@ -73,83 +55,108 @@ func TestWithQueueSize(t *testing.T) {
 
 func TestWithMaxPages(t *testing.T) {
 	tests := []struct {
-		name        string
-		input       int
-		expectRet   string
-		expectedVal int
+		name  string
+		input uint64
+		err   error
 	}{
 		{
-			name:        "page size can set",
-			input:       4,
-			expectRet:   "",
-			expectedVal: 4,
+			name:  "page size can set",
+			input: 4,
+			err:   nil,
 		},
 		{
-			name:        "page size cannot set with less than 2",
-			input:       1,
-			expectRet:   "maxPages must be set to greater than 1",
-			expectedVal: 0, // nil
+			name:  "page size cannot set with less than 2",
+			input: 1,
+			err:   ErrQueueOption,
 		},
 	}
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			var options options
-			f := WithMaxPages(tt.input)
-			actualRet := f(&options)
-			if actualRet == nil {
-				if tt.expectRet != "" {
-					t.Fatalf("Failed test: %s, expectRet: %v, actualRet: %v", tt.name, tt.expectRet, actualRet)
-				}
-			} else {
-				if !strings.Contains(actualRet.Error(), tt.expectRet) {
-					t.Fatalf("Failed test: %s, expectRet: %v, actualRet: %v", tt.name, tt.expectRet, actualRet)
-				}
+			f := WithMaxPage(tt.input)
+			err := f(&options)
+			if !errors.Is(err, tt.err) {
+				t.Errorf("failed test: got is not equal err, %v, %v", err, tt.err)
 			}
-			actualVal := options.maxPages
-			if actualVal == nil {
-				if tt.expectedVal != 0 {
-					t.Fatalf("Failed test: %s, expectedVal: %v, actualVal: %v", tt.name, tt.expectedVal, actualVal)
-				}
-			} else {
-				if *actualVal != tt.expectedVal {
-					t.Fatalf("Failed test: %s, expectedVal: %v, actualVal: %v", tt.name, tt.expectedVal, actualVal)
-				}
+			if err == nil && *options.maxPage != tt.input {
+				t.Errorf("failed test: got is nil, %v", err)
 			}
 		})
 	}
 }
 
 func TestWithEncoder(t *testing.T) {
-	var input func(v any) ([]byte, error) = func(v any) ([]byte, error) { return []byte{0x00}, nil }
-	var expectRet error = nil
-	expectedVal := input
+	var encoder func(v any) ([]byte, error) = func(v any) ([]byte, error) { return []byte{0x00}, nil }
 
 	var options options
-	f := WithEncoder(input)
-	actualRet := f(&options)
-	if expectRet != actualRet {
-		t.Fatalf("Failed test: expectRet: %v, actualRet: %v", expectRet, actualRet)
+	f := WithEncoder(encoder)
+	err := f(&options)
+	if err != nil {
+		t.Errorf("failed test: got is nil, %v", err)
 	}
-	actualVal := options.encoder
-	if actualVal == nil {
-		t.Fatalf("Failed test: expectedVal: %p, actualVal: %p", expectedVal, actualVal)
+
+	if options.encoder == nil {
+		t.Errorf("failed test: got is not nil")
 	}
 }
 
 func TestWithDecoder(t *testing.T) {
-	var input func(data []byte, v any) error = func(data []byte, v any) error { return nil }
-	var expectRet error = nil
-	expectedVal := input
+	var decoder func(data []byte, v any) error = func(data []byte, v any) error { return nil }
 
 	var options options
-	f := WithDecoder(input)
-	actualRet := f(&options)
-	if expectRet != actualRet {
-		t.Fatalf("Failed test: expectRet: %v, actualRet: %v", expectRet, actualRet)
+	f := WithDecoder(decoder)
+	err := f(&options)
+	if err != nil {
+		t.Errorf("failed test: got is nil, %v", err)
 	}
-	actualVal := options.decoder
-	if actualVal == nil {
-		t.Fatalf("Failed test: expectedVal: %p, actualVal: %p", expectedVal, actualVal)
+
+	if options.decoder == nil {
+		t.Errorf("failed test: got is not nil")
+	}
+}
+
+func TestWithGroupSize(t *testing.T) {
+	groupSize := 1
+	var options options
+	f := WithGroupSize(groupSize)
+	err := f(&options)
+
+	if err != nil {
+		t.Errorf("failed test: got is nil, %v", err)
+	}
+
+	if *options.groupSize != groupSize {
+		t.Errorf("failed test: got is equal groupSize, %d, %d", *options.groupSize, groupSize)
+	}
+}
+
+func TestWithQueueType(t *testing.T) {
+	tests := []struct {
+		name  string
+		input QueueType
+	}{
+		{
+			name:  "SPSC",
+			input: SPSC,
+		},
+		{
+			name:  "MPSC",
+			input: MPSC,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			var options options
+			f := WithQueueType(tt.input)
+			err := f(&options)
+			if err != nil {
+				t.Errorf("failed test: got is nil, %v", err)
+			}
+			if *options.queueType != tt.input {
+				t.Errorf("failed test: got is equal input, %v, %v", *options.queueType, tt.input)
+			}
+		})
 	}
 }
