--- testdata/simple_queue/new_queue/ffq/queue.0
@@ -1,1 +1,0 @@
-[{"Value": 1}]
