--- testdata/utils/create_queue_dir/ffq/index
@@ -1,1 +1,0 @@
-0000012345
