--- testdata/utils/open_index_file/ffq/index
@@ -1,1 +1,0 @@
-0000012345
