--- testdata/utils/read_index/ffq/index_invalid_permission
@@ -1,1 +1,0 @@
-0000000000000012345
