--- testdata/utils/read_index/ffq/index_invalid_string
@@ -1,1 +1,0 @@
-invalid
